MicroServices/MeasurementDataService/requirements.txt
pandas==1.4.2
paho-mqtt==1.6.1